Rust syml protocol initial implementation
