+++ rust-tools/core/Cargo.toml
@@ -10,3 +10,4 @@ edition = "2018"
 home = "0.5.3"
 serde = {version = "1.0.125", features = ["derive"]}
 serde_json = "1.0.64"
+uuid = {version = "0.8.2", features = ["serde", "v4"]}
